+++ projekatTSEO/src/main/java/projekatTSEO/repository/KorisnikRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import projekatTSEO.model.Korisnik;
 
-public interface KorisnikRepository extends JpaRepository<Korisnik, Integer> {
+public interface KorisnikRepository extends JpaRepository<Korisnik, Long> {
 
 	Korisnik findByKorisnickoIme(String korisnickoIme);
 }
