+++ projekatTSEO/src/main/java/projekatTSEO/repository/PredavanjePredmetaRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import projekatTSEO.model.PredavanjePredmeta;
 import projekatTSEO.model.Profesor;
 
-public interface PredavanjePredmetaRepository extends JpaRepository<PredavanjePredmeta, Integer>{
+public interface PredavanjePredmetaRepository extends JpaRepository<PredavanjePredmeta, Long>{
 
 	List<PredavanjePredmeta> findByProfesor(Profesor profesor);	
 }
