+++ projekatTSEO/src/main/java/projekatTSEO/repository/PredmetRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import projekatTSEO.model.Predmet;
 
-public interface PredmetRepository extends JpaRepository<Predmet, Integer> {
+public interface PredmetRepository extends JpaRepository<Predmet, Long> {
 	
 	Predmet findByNaziv (String naziv);
